route: add basic route validation
A route is created using the default certificates for routes for
the outside world. It then uses the default CA for the service
it is associated to (service serving CA) to reencrypt the communication
on the inside.

The route gets a default external hostname: <route-name>[-<namespace>].
Currently, that leads to it being called
openshift-osin-openshift-osin.apps.CLUSTER_NAME.CLUSTER_DOMAIN