Move OAuth metadata publishing right after route handling
As we don't yet have a cluster operator deployed by the CVO,
we need to make sure that the OAuth metadata is published as
soon as we can so that it is reflected in the API server for
the other components.

Once we get the cluster operator deployed by CVO, this can be
done at any time as CVO will need to wait for us to report
availability. Make sure to keep in mind that everything has
to be in place to report available status, then!