Add defaulting for operatorv1.Authentication
This change makes it so that we render an empty
operatorv1.Authentication on start.  This is then defaulted to the
managed state at runtime.  A missing Authentication resource is
treated as the same as an empty one.

The defaults are stored once in the defaultCopyAuthenticationFunc
function (they are not duplicated in any manifests).  They are also
applied regardless of if the resource exists in the API (any unset
field means "give me the default").

Signed-off-by: Monis Khan <mkhan@redhat.com>