OpenID: add discovery via /.well-known/openid-configuration
This change adds support for discovering OpenID configuration via
the issuer's /.well-known/openid-configuration endpoint.  This
reduces the config that is required (and the associated chance of
configuration errors) while preserving the same feature set.

Temporary fallback code was added to respect the old urls struct.
This allows an admin to migrate configuration and prevents upgrades
from breaking.

Signed-off-by: Monis Khan <mkhan@redhat.com>