redeploy on operatorConfig.spec changes
Since we no longer track the resource version of operatorConfig in
the OAuth server's deployment, a change to operatorConfig.spec will
not cause ApplyDeployment to issue an update.  Because we track the
generation in operatorConfig.status, we can compare it to determine
if a force redeploy is required.  This is effectively the same as if
we used the operatorConfig.generation as its "resource version."

Signed-off-by: Monis Khan <mkhan@redhat.com>