rearrange(manifests): rename files for proper naming in release bundle
The release build process adds the "cluster-authentication-operator" prefix
if no runlevel is specified (i.e. if the filename does not have the `000_[0-9]+_`
 prefix). That means our release manifest files end up with it twice. So like,
 `manifests/cluster-authentication-operator_00_roles.yaml` ends up as
`0000_70_cluster-authentication-operator_cluster-authentication-operator_00_roles.yaml`
in the release bundle.

This PR standardizes our manifest file names with the other cluster operators
and dedups the "cluster-authentication-operator" labeling in the relase files.