Correctly determine operand version
This change updates the operator deployment to include the operand
version as an environment variable.  Since we encode the operator
deployment RV into the operand deployment, we correctly gate on
changing the version until the new version has rolled out.

The VersionForOperand function uses a config map with a version to
image mapping to determine the operand version.  It does this by
reversing the mapping (image to version).  This has a serious flaw
in that it assumes that all versions point to distinct images
(which is not the case since an image may not change during a
release).  Since map iteration order is undefined, this can lead to
the operand version flapping back and forth as the reversed map
changes based on which image value was written last.

Remove version-mapping config map as it is no longer used.

All environment variable names and values are now stored as consts
and vars, respectively (the values are effectively runtime consts as
they cannot change once the operator is running).

Signed-off-by: Monis Khan <mkhan@redhat.com>