Make oauth-openshift.<ingress_domain> the permanent issuer of the OAuth server
We need a permanently stable issuer URL for the OAuth server - both
in terms of the OAuth spec and so that OAuth callbacks from OAuth
IDPs continue to work indefinitely (we have no way to reconcile the
external IDP on a change).

oauth-openshift is the shortest namespaced value we can use (we do
not want to use a value that a customer may want to use).  This is a
best effort attempt to make the external API "nice" to use.

To keep things consistent, everything that was previously exposed as
"integrated-oauth-server" is now "oauth-openshift" (this does make
us slightly inconsistent with the top level Authentication config
and the binary of the OAuth server - we can at least update the
latter in the future for consistency).

Bug 1706894

Signed-off-by: Monis Khan <mkhan@redhat.com>