Honor the None authentication type and WebhookTokenAuthenticators
When the None authentication type is set, we shouldn't be adding the
reference to the IntegratedOAuthMetadata config map. This way the
kube-apiserver-operator will nt attempt to track the configmap.

It also takes into account if the webhooktokenauthenticators are set;
because, if they're not, you'll end up in an environment where you can't
log in.