Add strict validation for router certs
The ingress operator performs no validation on the certs it is given
by the user.  Thus it can easily pass us nonsensical data.  We guard
against it and provide useful error messages.

This change also detects if the ingress operator ever wedges and
fails to keep the router-certs updated with the correct certs on an
ingress.config domain plus cert change.

Signed-off-by: Monis Khan <mkhan@redhat.com>