Make SNI output stable when multiple ingress domains are configured
When multiple ingress domains are configured, the router-certs
secret will have a key for each domain.  When we build the named
certificates slice to pass to the OAuth server via its CLI config,
we need to guarantee that the output is stable.  Otherwise the OAuth
server will be consistently redeployed since the config hash will be
different each time.

Since router-certs is a map, we need to iterate over it in a
deterministic manner to make the output stable.

Signed-off-by: Monis Khan <mkhan@redhat.com>