Ensure primary controller only updates conditions it owns
Previously the primary auth controller set all condition types when
performing an update. With the addition of the ingress state and
router certs controllers, this behavior resulted in status message
flapping as the primary controller raced with the new controllers. The
fix is having the primary controller avoid setting the condition types
managed by the other controllers.