bug 1871890: add current profile annotations to CVO manifests
Cluster profiles are a way to support different deployment models for OpenShift clusters. A profile is an identifier that the Cluster Version Operator uses to determine which manifests to apply. Operators can be excluded completely or can have different manifests for each supported profile.

To support above, a deployment model in which not all operators rendered by the CVO by default is needed. The use case includes IBM Public Cloud, in which a hosted control plane is used. Potentially it can also be used for Code Ready Containers.

The following annotation may be used to include manifests for a given profile:

include.release.openshift.io/[identifier]=true

This would make the CVO render this manifest only when CLUSTER_PROFILE=[identifier] has been specified.