expose pod information to oauth-apiserver container via env variables
the pod name and the namespace are used by the termination code to properly record events.

before the change the termination events were recorded to the default namespace and weren't associated with any pod.
after the change the termination evens are recorded in the openshift-oauth-apiserver namespace and are bound to oauth-apiserver pods.