only configure webhook authenticators when oauth-apiservers are ready during upgrade
If we configure the webhook authenticators too early during an upgrade,
we might end up with apiservers that are incapable of handling
TokenReviews. Make sure that the version of the oauth-apiserver payload
in the version getter global to the operator is current if it is set.
If it is unset, this should not be an upgrade.