observe api-audiences for the oauth-apiserver
When the kube-apiserver receives a token validation request,
that request is audience-constrained since we configure the SA
issuer and audiences. This means that kube-apiserver also checks
the audiences returned from a TokenReview during webhook token
authentication.

The oauth-apiserver by default uses the https://kubernetes.default.svc
audiences. If authentication.spec.serviceAccountIssuer is set however,
the kube-apiserver will expect different audiences from the default.
If the oauth-apiserver does not adjust the TokenReview endpoint,
users won't be able to validate their OAuthAccessTokens as the
audience returned by the oauth-apiserver in a TokenReview response
would differ.