manifests, bindata: explicitely set runAsUser for operator and operand
Both cluster-authentication-operator and openshift-oauth-apiserver need to run as root in order to be able to overwrite /etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem.

This is implied from picking the default anyuid SCC.
However, if user specifies a more restricted SCC specifying MustRunAsRange with higher or equal priority than anyuid,
that one will be picked causing both processes to be started as non-root.

This fixes it by specifying runAsUser: 0 explicitely.